refactor: Update fetchData.js and antTableHelper.js
This commit updates the fetchData.js and antTableHelper.js files. It adds a new function fetchFilteredData to handle fetching filtered data from the server. The function makes a POST request to the '/data/filter' endpoint with the provided parameters and updates the state variables accordingly. In antTableHelper.js, the handleFilterDropdownVisibleChange function is added to generate filter options for a given collection and field. It makes a POST request to the '/data/generateFilters' endpoint and updates the filters state variable.