refactor: Update UserList.js to use more descriptive labels for radio buttons
This commit updates the UserList.js file to use more descriptive labels for the radio buttons in the UserList component. Instead of "Users" and "Engagement", the labels are changed to "Registered Users" and "Registered Users + Leads" respectively. This change improves the clarity and usability of the radio buttons, providing a better understanding of the options available to the user.