feat: Update UserEngagement table to include lastReached column with DatePicker
The UserEngagement table has been updated to include a new column called lastReached. This column displays the date and time when the user was last reached. It is represented by a DatePicker component, allowing users to easily select the date. This enhancement improves the functionality and provides a better user experience.